Reorganization to put project into GitHub.
- New structure for the project folders and files;
- Git ignore file created;
- Batch script files changed to fit new folder structure.
