+++ Compile.bat
@@ -1,4 +1,4 @@
-windres resources.rc -o resources.o
-gcc main.c resources.o -l gdi32 -mwindows -o Snake.exe
+windres src\resources.rc -o obj\resources.o
+gcc src\main.c obj\resources.o -l gdi32 -mwindows -o bin\Snake.exe
 echo %ERRORLEVEL%
 pause
