+++ src/resources.h
@@ -1,0 +1,26 @@
+//Icons
+#define IDI_ICON256             101
+#define IDI_ICON48              102
+#define IDI_ICON32              103
+#define IDI_ICON16              104
+
+//Menu
+#define IDM_MENU                40001
+#define IDM_FILE_NEW            40002
+#define IDM_FILE_PREFERENCES    40003
+#define IDM_FILE_EXIT           40004
+#define IDM_ABOUT               40005
+
+//Keyboard Accelerator
+#define IDK_ACCELERATORS        200
+
+//Dialog Box
+#define IDD_ABOUTBOX            300
+#define IDD_PREFERENCES         301
+
+//Controls
+#define IDC_STATIC              -1
+#define IDC_FIELD_WIDTH         400
+#define IDC_FIELD_HEIGHT        401
+#define IDC_SPEED               402
+#define IDC_PASS_THROUGH_WALLS  403
